feat: BaseModel的id字段改为int64

+++ server/biz/model/BaseModel.go
@@ -3,7 +3,7 @@ package model
 import "gorm.io/plugin/soft_delete"
 
 type BaseModel struct {
-	ID        uint                  `gorm:"primarykey" json:"id"`
+	ID        int64                 `gorm:"primarykey" json:"id"`
 	CreatedAt int64                 `gorm:"autoCreateTime:milli" json:"created_at"`
 	UpdatedAt int64                 `gorm:"autoUpdateTime:milli" json:"updated_at"`
 	DeletedAt soft_delete.DeletedAt `gorm:"softDelete:milli" json:"deleted_at"`
